cpuspeed: add NDIS v4 speed patch suitable for WFW 3.11
WFW 3.11 uses a similar timing loop as Win9x in NDIS.386 and so will hang
during startup on a blank screen on a machine with a fast CPU.

Fortunately due to its shared history it is possible to take the existing
NDIS v1/v3 source and adapt it for WFW 3.11. Add the updated source as a
new NDIS v4 source, include it PATCH_CPU_SPEED_NDIS_ALL, and update files.c
to patch NDIS.386 if detected.

Resolves: https://github.com/JHRobotics/patcher9x/issues/23